fix: snapshots tests that are using TimeProvider
