+++ quick-scan/settings/SettingsViewControllerFactory.swift
@@ -4,9 +4,10 @@ import UIKit
 final class SettingsViewControllerFactory {
     class func create(
         userDefaults: UserDefaults = UserDefaults.standard,
-        version: Versioning = BundledVersion()
+        version: Versioning = BundledVersion(),
+        timeProvider: TimeProvider = FoundationTimeProvider()
     ) -> UIViewController {
-        let view = SettingsView(version: version).defaultAppStorage(userDefaults)
+        let view = SettingsView(version: version, timeProvider: timeProvider).defaultAppStorage(userDefaults)
         let viewController = UIHostingController(rootView: view)
         viewController.title = Strings.Settings.title
         return viewController
